LayoutTests/platform/glib/http/tests/websocket/tests/hybi/imported/blink/permessage-deflate-unsolicited-negotiation-response-expected.txt

Test whether WebSocket ignores unsolicited negotiation response.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS onopen
FAIL ws.extensions should be permessage-deflate; server_max_window_bits=15; server_no_context_takeover. Was permessage-deflate; server_no_context_takeover; server_max_window_bits=15.
PASS event.wasClean is true
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
